Fixed struct-parsing file
Removed all unused includes. Also, C doesn't support bool type until
C99, which is not the default for all compilers. So added a feature
require in the CMakeLists.txt to ensure that we use the last C standard
(C11). Finally, when using bool, we should include the stdbool.h which
define it and it's replacement when unsupported.

Change-Id: Iecca1458cb77577418a9d0af7d4f43d1b49ddd70
Signed-off-by: Loïc Collignon <loic.collignon@iot.bzh>